clean: optional dependencies modules
