+++ lib/algoliax/resources/object.ex
@@ -1,9 +1,6 @@
 defmodule Algoliax.Resources.Object do
   @moduledoc false
 
-  import Algoliax, only: [import_if_loaded?: 1]
-  import_if_loaded?(Ecto.Query)
-
   import Algoliax.Utils,
     only: [
       index_name: 2,
@@ -64,66 +61,6 @@ defmodule Algoliax.Resources.Object do
     module
     |> index_name(settings)
     |> Requests.delete_object(object)
-  end
-
-  if Code.ensure_loaded?(Ecto) do
-    def reindex(module, settings, index_attributes, query, opts \\ [])
-
-    def reindex(module, settings, index_attributes, %Ecto.Query{} = query, opts) do
-      repo = Algoliax.Utils.repo(settings)
-
-      Algoliax.Utils.find_in_batches(repo, query, 0, settings, fn batch ->
-        save_objects(module, settings, batch, index_attributes, opts)
-      end)
-    end
-
-    def reindex(module, settings, index_attributes, _, opts) do
-      repo = Algoliax.Utils.repo(settings)
-
-      modules =
-        case Algoliax.Utils.schemas(settings) do
-          [_ | _] = schemas ->
-            schemas
-
-          _ ->
-            [module]
-        end
-
-      modules
-      |> Enum.each(fn mod ->
-        query = from(m in mod)
-
-        Algoliax.Utils.find_in_batches(repo, query, 0, settings, fn batch ->
-          save_objects(module, settings, batch, index_attributes, opts)
-        end)
-      end)
-
-      {:ok, :completed}
-    end
-
-    def reindex_atomic(module, settings, index_attributes) do
-      Algoliax.Utils.repo(settings)
-
-      Index.ensure_settings(module, settings)
-
-      index_name = index_name(module, settings)
-      tmp_index_name = :"#{index_name}.tmp"
-      tmp_settings = Keyword.put(settings, :index_name, tmp_index_name)
-
-      Algoliax.SettingsStore.start_reindexing(index_name)
-
-      reindex(module, tmp_settings, index_attributes, nil)
-
-      Requests.move_index(tmp_index_name, %{
-        operation: "move",
-        destination: "#{index_name}"
-      })
-
-      Algoliax.SettingsStore.delete_settings(tmp_index_name)
-      Algoliax.SettingsStore.stop_reindexing(index_name)
-
-      {:ok, :completed}
-    end
   end
 
   defp build_batch_object(module, settings, model, attributes, "deleteObject" = action) do
